+++ demos/serverless/src/index.js
@@ -12,3 +12,16 @@ exports.handler = async (event, context, callback) => {
   response.body = fs.readFileSync('./index.html', {encoding: 'utf8'});
   callback(null, response);
 };
+
+exports.handlerV2 = async (event, context, callback) => {
+  var response = {
+    "statusCode": 200,
+    "headers": {
+      'Content-Type': 'text/html'
+    },
+    "body": '',
+    "isBase64Encoded": false
+  };
+  response.body = fs.readFileSync('./indexV2.html', {encoding: 'utf8'});
+  callback(null, response);
+};